Implement `AirColumnAccess` and `trace.LazyTable`
This implements `AirColumnAccess` which provides a mechanism for reading
a (potentially shifted) column value from a table.  In addition, this
adds an interface `trace.Table` for representing arbitrary tabular data
and a simple implementation of that interface, `trace.LazyTable`.  The
`trace.Table` interface needs to be fleshed out a fair bit more though.
For example, it does not support computed columns or hetrogeneous
columns (e.g. with differing length, etc).

Finally, the parser for AIR expressions does not support shifting in
form.  Therefore, tests cannot be added for that at this time.