Implement `ArrayTrace` and rename `trace` => `table`
This provides a concrete implementation of a `Trace` which simply stores
columns of raw data.  In addition, a `Schema` can now `Accept` a trace
(or not).  That corresponds to the previous `Check()` method which
essentially evaluates each constraint on the trace.

At this stage, there are still some tests failing.  These are related to
situations where a trace cannot be rejected at the HIR or MIR level
(i.e. because it contains malicious computed column information).  This
should be relatively easy to fix.