Implement Typed Columns
This adds typed columns down at the MIR level.  The logic here is
that, at the HIR level, we'll have another higher-level notion of a
column.  Specifically, to support: perspectives; vectored columns (i.e.
to support different field types).

This also adds the notion of a _range constraint_ at the AIR level.  Thus, a
typed column is translated as a regular column plus a range constraint.
This is now operational.