Connect `binfile` to `hir`
This puts in place a minimal connection from a `binfile` to an `hir`
representation.  This allows one to compile a set of corset constraints
using the Rust `corset` tool and then parse them into the HIR form.
However, there are a lot of features remaining to be implemented.  Thus,
it only really works for simple examples at this stage.