Support `first`/`last` specifiers and `ifnot`
This adds support for `first` or `last` specifiers on a vanishing
constraint.  This makes such a constraint local to a specific row of the
trace (either the first row, or the last row).

In addition, this adds a new conditional `ifnot A B` which essentially
compiles to `if A _ B`.  This new conditional can only have two
arguments.