Complete refactoring for HIR / MIR
This completes the refactoring for the HIR and MIR levels.  That seems
to be offering a useful improvement by, for example, removing duplicate
code.

I have also put in place a prelimary (and as yet incomplete) mechanism
for reporting warnings from constraints.  The intuition here is that a
property assertion failure should not be the reason a trace is rejected
(i.e. because the prover cannot do that).  However, at the same time, an
accepted trace which fails a property assertion should be flagged in
some way.