Implement `Testable` interface
This subsumes `Evaluable` in situations where we want an expression
which can be (kind of) evaluated, but isn't actually `Evaluable`.
Specifically, expressions at the HIR level are not `Evaluable` because
they can return multiple values.  Despite this, we can consider whether or
not a given expression evaluates to zero on a given row.