HIR List Evaluation now Operational
List evaluation was previously broken because we required that HIR
expressions implemented `Evaluable`.  However, this was impossible for
`hir.List` expressions because they can return more than one evaluation
point.

The solution is to avoid requiring `hir.Expr` to implement `Evaluable`.
Instead, we have the weaker requirement that `hir.Expr` implements
`Testable`.  The latter can hide the fact that we have multiple
evaluation points.  Overall, this seems to work quite well.

A small number of tests specifically aimed at blocks are included in
this PR.