Add SortedPermutation to HIR
This adds support for parsing and storing sorted permutations.  This
also adds more useful permutation tests.  Constraints for enforcing
sorting and generated at the HIR and MIR level, but not yet at the MIR
level.  Computations for expanding sorted permutations at all levels are
operational.

This additionally splits the AIR pipeline into constraints versus
computations.  The former are restricted to a minimal set of primitives,
whilst the latter are essentially open ended.

Finally, this adds proper support for dealing with expanded test traces.
Specifically, when an expanded trace is detected then trace expansion at
the AIR level is not performed.