Implement wide range constraints
This implements so-called "wide range constraints" using byte
decomposition.  In essence, when a column is declared with a e.g. a type
`u32` then this automatically constructs a byte decomposition to ensure
values of the given type fit.

One issue, at this stage, is that there is no way to distinguish
checked from unchecked column types.