Insert Initial Padding Row
The initial padding row is inserted during trace expansion by default
now.  This could be extended with spillage.  This required update many
of the tests to ensure they are "compatible" with an initial row of
padding.  This makes sense as, in fact, any constraint set must be
compatible with an arbitrary amount of padding.