Account for Spillage
This updates the system to enable correct accounting of spillage.  At
this stage, there remains an outstanding problem related to the handling
of sorted permutations.  In particular, the padding values are not
determined correctly.  The issue seems to be that, for a negative
permutation sort, we actually want to pad at the back (i.e. not the
front).

I think the next step is to investigage what Corset does.  Perhaps this
is actually where it uses back padding.