Partial fix for spillage issue.
This puts in place a partial fix for the spillage issue.  The intuition
here is that we have a sign for padding, and that padding can happen on
the front or the back.

However, at the moment, it is broken because when padding is applied the
height of the trace is not being updated.  The best solution will be to
add a method `Pad()` to `Trace` which accepts an iterator over the signs
of the column.  This can be used to properly pad the columns in one go.