Resolve issues around negative permutation sorts
There are some genuinely hard problems with negative permutation sorts.
The current solution is simply to prohibit permutation sorts whose
"index column" (i.e. leftmost column) is negatively sorted.  There is an
additional issue that traces which have a `0` for this column prior to
expansion will (most likely) be rejected.