Support Bounds Computation
This puts in place a proper notion of a bounded computation, which
replaces the previously awkward notion of `MaxShift`.

This also requires that every column in a trace is given a
`padding` value which is used when padding that column.  The value for
input columns is always `0` otherwise it is computed during trace
expansion.  Out-of-bounds accesses on trace columns then return the
padding value.