Schema modules partially working
The basics of schema-level modules are now in place, and we have two
tests added for them.  The outstanding problem is that of alignment.
Specifically, trace and schema module alignment.  Its not clear to me
this should be significant.  However, there is an issue whereby a schema
module index is being used (during alignment) to determine the correct
column index.