Fix Misalignment
This fixes an issue with alignment whereby modules exist in the schema,
but not in the trace.  This can arise in valid circumstances.  For
example, when a module is declared which doesn't contain any columns.
Another situation is the prelude module, which is always declared
regardless of whether or not any columns are declared.