Support Evaluation Context with Length Multiplier
The notion of an evaluation context has morhped from being just that of
the enclosing module, to also include the length multiplier.  In effect,
modules are divided up into different demarcations based on the length
multiplier.  Thus, for example, every expression must have a unique
enclosing module and length multiplier.