Propagate evaluation context throughout
This completes propagating the new type `Context` throughout the code
base.  In the end, `Context` was moved into the `trace` package because
it is used in both `schema` and `trace` and this was necessary to avoid
a cyclic dependency.  That seems reasonable given that the evaluation
context is inherently about evaluation :)