Preliminary fix for Perpsectives
This is a preliminary fix for managing registers versus columns for
perspectives.  I actually don't think this is quite ready yet.  The
problem is that the `bin` file allocates columns for permutations.

However, its not at all clear to me that the allocated register for a
sorted permutation will align with the `Assignment` created for a
permutation.  Furthermore, I believe we are currently creating two
columns for each target permutation column.  One is being added as a
data column, whilst the other is being added as an Assignment.