Fix #202 using custom `HashSet` implementation
This provides a custom `HashSet` implementation for use when checking
lookup constraints.  This improves the time complexity for
`LookupConstraint.Accepts()` from `O(n^2)` to approx `O(n)`.

The reason for using custom `HashSet` implementation is simply to ensure
that collisions are handled correctly.  Existing `HashSet`
implementations (e.g. hashicorp) do not safely support collisions.  That
is, they assume the hash function never collides for two different
items.  Since I do not want to make that assumption, I have implemented
my own variation.  This is a true hashtable in the sense that it uses
buckets to handle collisions, and requires an extended `Hasher`
interface (compared with ithat in e.g. hashicorp's `go-set`).
Specifically, the `Hasher` interface includes a notion of equality.

A reasonable number of tests have been added for this `HashSet`
implementation.