Support Sequential ParExec
This puts in place support for using a sequential ParExec
implementation.  This is a first step towards parallelising trace
expansion.  Specifically, there are two outstanding issues to resolve:
firstly, we must be able to add expanded columns in arbitrary order;
secondly, we need to actually make `ParExec` operate in parallel (which
make require making trace thread-safe or something).

The issue around the order in which expanded columns are added to the
trace is problematic as the current design essentially expects them to
be added in a specific order.  Instead, we need assignments and input
columns to know what their target column index is.  Then, we expanding
the trace, they can specify this.