Support Assignment.ComputedColumn
This replaces the previous `Assignment.ExpandTrace` method which
directly added columns to the trace.  Instead, computed columns are
simply returned and its left upto the caller to add them to the trace.
For now, I've put in place some hacky glue code which does this ... and
passes the tests.

The next step I think is to package everything to do with trace
expansion into a trace builder.  That will provide a single point where
things will break related to the trace API.  In particular, the goal is
to dramatically simplify the notion of a trace to just a pair of
([]Module,[]Column) arrays.