Refactor Trace API
This puts through a significant and wide-reaching refactoring of the
Trace API.  Amongst other things, this dramatically simplifies the API.
Furthermore, it changes the emphasis around trace construction
particuarly with respect to trace expansion.  It is not longer the case
that trace expansion must occurr in a predetermined order.  Rather,
assignments can be expanded in whatever order is required (provided
dependencies are met).  This therefore, now enables the possibility of
parallel trace expansion.