Support Parallel Trace Expansion
This puts in place a relatively straightfoward parallel algorithm for
trace expansion using go routines.  Given all the previous refactoring
to enable trace expansion to occurr out-of-order, this was relatively
easy to implement.  Performance improvements are not substantial, but
this is perhaps to be expected given that checking is more compute
heavy.