Support spillage on a per-module basis
This computes spillage on a module-by-module basis, rather than
computing one conservative value that covers all modules.  This means
the spillage computation is more precise, but also slightly more
involved.  To make this work, `Assignment.Context()` was added so that
we can identify the context in which an assignment is declared.