Fix debug output
This fixes the output for the `debug` command.  Specifically, it now
prints variable names (rather than cryptic variable indices).  The
entire mechanism for printing expressions has been replaced with a
generic mechanism for converting into lisp expressions (sexp.SExp).
This should allow, in the future, the option to format code.

In addition, this fixes various issues with the --stats option which was
missing various constraints.