Support `entropy` column summariser
This adds support for showing the "entropy" of a column.  That is a
count of how much the column changes between each line.  High entropy
signals few changes between lines (e.g. all 1s as 100% entropy).

This also refactors the column summariser process to allow selection of
which summaries to show.