Tweak test command
This tweaks the test command so that it doesn't treat property
assertions as "absolute truth".  Rather, it attempts to find traces
which are accepted but where the property assertions don't hold.

At this stage, the system is fairly simplistic.  We need to think more
carefully about how to generate the traces.  This includes avoiding the
brute-force enumeration currently in play.  Likewise, thinking more
carefully about the "pool" of values used for testing.