Rework scope resolution
This reworks the way in which scopes are resolved to avoid using the
Environment until the very last moment in the compilation process.  This
just means we don't have to get too concerned about specific column
identifiers until it really matters.  It also means we can retain more
detailed meta-information, allowing better analysis.