feat: implement defconst (#419)
* Add tests for `(defconst ...)` declarations

This adds a range of tests based on existing tests for the `(defconst
...)` declarartion.  They are all failing of course :)

* Update parser for defconst

This updates the parser to parse defconst declarations.  This also adds
a number of invalid tests as well.

* Constants are mostly working

Constants are now mostly working, but there are a few edge cases to work
through.

* Final Fixes

The final fixes are related to generalising exponentiation and shifting, as well as proper checking of identifiers during parsing.