feat: support `reduce` (#445)
* Add test cases for reduce

* Support reduce operator

This adds support for the reduce operator, which performsing folding
over binary functions.

* Support intrinsics

This adds support for intrinsics which are needed for reductions
(amongst other things).  Specifically, they are required for reductions
which operate over built-in operators, such as "+", "-", etc.  At this
stage, only a few intrinsics have been added and I'm anticipating that
more will be required at some point.