feat: specify error for invalid tests (#467)
* Trying to extract errors

* Embed errors in test files

This puts in place the mechanism for embedding error messages into the
invalid tests themselves.  That way, anything which changes what errors
are reported results in a failing test.

* Fixes for error checking

This puts through a few more fixes for the error checking mechanism, and
adds errors for basic and constant tests.  There's a fair way to go yet
though.

* Updated all invalid tests

This updates all the existing invalid tests to include specific information
about the expected errors.  Overall, this seems to be working well.