fix: mxp test (#483)
* Support slicing traces

This puts in place proper support for slicing traces using the "trace"
command.

* Add failing test case

This adds a minimal failing test case extracted from the mxp trace /
lisp.  It looks rather like the loop "for i [5]" is the problem.

* Fix for loop with unit bound

This puts in place a very simple fix for for loops with a range of the
form "[n]", rather than the more usual "[n:m]".  It was a simple logic
error.