feat: update shf test (#491)
Update shf test

This updates the shf test itself, rather than changing the go-corset
implementation.  This is because, at this time, it seems like the more
sensible option.  However, we may need to revisit this in the future if
it ends up forcing lots of changes in the full constraint set.