Parsing and (incorrectly) translating perspectives
Parsing of perspectives is working now, and a translation is being made.
This checks of the use of perspectives within constraints, and
expressions.

At this time, perspectives are now be translated correctly.  However,
this translation remains naive because there is no sharing of columns.
That is, no register allocation is being performed!