Split Registers from Columns
This splits the concepts of (HIR) registers away from Corset columns.
Specifically, the GlobalEnvironment becomes responsible for allocating
Corset columns to HIR registers.  At the moment, this allocation is
essentially the same as it was before.  However, it is now possible to
do something more sophisticated here.

NOTE: it would be worth investigating the use of perspective columns in
lookups, interleavings and permutations.  Its not clear which of these
would or could support the use of perspective columns.