Implement Register Allocation
This puts in place the notion of a "greedy allocator".  This is a
register allocator which works in a specific "greedy" fashion.  However,
it is actually quite flexible in terms of how its used and, hence, it
can be used to implement a wide variety of allocators with differing
levels of efficiency.

Register allocation is now feature complete for supporting the original
allocator.  However, more advanced allocators would require further
implementation work.  At this stage, it remains to add a number of
suitable perspective tests.