feat: support corset perspectives (#500)
* Parsing and (incorrectly) translating perspectives

Parsing of perspectives is working now, and a translation is being made.
This checks of the use of perspectives within constraints, and
expressions.

At this time, perspectives are now be translated correctly.  However,
this translation remains naive because there is no sharing of columns.
That is, no register allocation is being performed!

* Split Registers from Columns

This splits the concepts of (HIR) registers away from Corset columns.
Specifically, the GlobalEnvironment becomes responsible for allocating
Corset columns to HIR registers.

This puts in place some preliminary mechanism for supporting arbitrary
register allocation algorithms.  At this stage, there is an issue around
the @prove tag.  Its somewhat unclear how best to support this.

This puts through a simple fix relating to the handling of array columns
being translated into registers.

* Implement Register Allocation

This puts in place the notion of a "greedy allocator".  This is a
register allocator which works in a specific "greedy" fashion.  However,
it is actually quite flexible in terms of how its used and, hence, it
can be used to implement a wide variety of allocators with differing
levels of efficiency.

Register allocation is now feature complete for supporting the original
allocator.  However, more advanced allocators would require further
implementation work.  At this stage, it remains to add a number of
suitable perspective tests.

* Add various tests