feat: support perspective name qualification (#504)
* Add initial tests

This adds some initial tests for perspective qualified notation.

* use paths intead of explicit qualifiers

This updates the implementation to use the notion of a "path" rather
than a trio of (module, perspective, name).  This should future proof us
for recursive submodules, and also simply helps ensure things actually
make sense.

* Replace `GlobalScope` and `ModuleScope`

This replaces these two constructs with a single new construct (called,
helpfully, `ModuleScope`).  The difference is that the new construct can
describe nested trees, thus making it suitable for handling perspectives
(and later submodules).

At this stage, register allocation is now working again with
perspectives.  I've chosen to retain backwards compatibility in terms of
naming with the original corset.  The issue is around the naming of
registers which arise from coalescing perspective columns.  The names
currently do not include the perspective name, and this means they can
potentially clash.  However, clashing is not a critical issue at this
time, since it will always result in the generated Trace.java file
containing a syntax error.