Add compile cli cmd
This adds a "compile" command to the CLI.  This is where compilation
will be controlled.  A notion of "legacy" versus "non-legacy" is patched
in.

This goal here is to retain the ability to read legacy bin files using
the original format (for now).  However, writing them is not really
supported at this time.