feat: binary constraints format (#514)
* Add compile cli cmd

This adds a "compile" command to the CLI.  This is where compilation
will be controlled.  A notion of "legacy" versus "non-legacy" is patched
in.

This goal here is to retain the ability to read legacy bin files using
the original format (for now).  However, writing them is not really
supported at this time.

* support binary encoding / decoding

This adds support for binary encoding / decoding of HIR schema's via the
gob format.  To help ensure consistency between compiled source files
and encoded binary files, all tests are additionally run through a
filter which encodes and then decodes the file before running the test.

This also makes a breaking constraint by using an option to implement
the vanishing domain, instead of a pointer.  This seems to serialise
better.