Initial implementqation of let
This adds a number of test cases for the `let` expression.  Most of
these tests are adaptions of existing tests.  Many of these tests are
passing at this stage.

There is a problem is related to the fact that type checking and name
resolution are done in one step.  This is causing an issue because we
end up (incorrectly) reresolving the body of an invoked function.  In
the presence of let expressions, this ultimately can break.