feat: support let syntax (#522)
* Initial implementqation of let

This adds a number of test cases for the `let` expression.  Most of
these tests are adaptions of existing tests.  Many of these tests are
passing at this stage.

There is a problem is related to the fact that type checking and name
resolution are done in one step.  This is causing an issue because we
end up (incorrectly) reresolving the body of an invoked function.  In
the presence of let expressions, this ultimately can break.

* Split name resolution from type checking

This puts through a fairly major change to the compilation pipeline by
splitting out name resolution from typing.  This had been previously
noted as something which could be done, and handling let expressions had
forced the issue.  Overall, the split seems to work out well enough.

* Allow type checking even if resolution fails

This allows the type checker to run even when resolution has produced
some errors.  This just means we get more errors being reported for
incorrect code, and more closely reflects how error reporting was
previously working.