feat: support `defcompute` (#547)
* support initial parser for defcompute

This puts in place an initial parser for defcompute declarations.

* Working on resolving defcomputed decls

* ready to translate defcomputed

Everything is in place to translate defcomputed declarations.  This will
require some (limited) support at the HIR level.

* defcomputed now being compiled

The defcomputed construct is now being compiled all the way down to the
AIR level.  At this stage, however, there are no actual native function
implementations.

* compute tests now passing

There is now a single native function for testing purposes (id).  This
is passing tests.

* Support for filter native function

This is another test native function which might be useful for
something.  Mostly its just for testing.

* Support filter function

The native filter function provides just another example of a
defcompute declaration.  It is really another test function, but it is
still useful.

* Fix broken tests

A number of tests defined an "id" function which is now redefined as a
native.  Therefore, we had a name clash.