feat: optimise normalisation over binary values
For a normalisation `(~ X)` arises for a binary column `X` (or
expression resulting in a binary value) then, in fact, no normalisation
is actually required.  Instead, we can just return `X` since it is
already normalised.  This is a useful optimisation since it will save an
inverse column, along with two constraints.