feat: optimise normalisation over binary values (#600)
* feat: optimise normalisation over binary values

For a normalisation `(~ X)` arises for a binary column `X` (or
expression resulting in a binary value) then, in fact, no normalisation
is actually required.  Instead, we can just return `X` since it is
already normalised.  This is a useful optimisation since it will save an
inverse column, along with two constraints.

* fix for broken test

Test_Type_05 was broken because it had a binary BIT with value 2 in its
accepts data.