fix for defcompute typing
This puts in place a straightforward fix for typing defcompute
constructs.  The problem was that, at the corset level, the type of the
target column was being extracted from the wrong input column.
Amazingly, the tests all still passed because some of the input columns
and unrestrictive (field) types.  Have added tests which do fail with
this error.