fix: incorrect type for defcompute (#614)
* move trace validation into `schema.Builder`

This just means trace validation can arise during testing as well.  Most
likely, we need to disable it for the reject tests.  This also adds some
tests for this issue.  However, it looks more like the issue is related
to permutation sorts, in fact.

* prevent validation for reject tests

This prevents trace validation from being applied to reject test.  The
reason for this is that such tests should fail for values which are
out-of-bounds on their own accord.

* fix for defcompute typing

This puts in place a straightforward fix for typing defcompute
constructs.  The problem was that, at the corset level, the type of the
target column was being extracted from the wrong input column.
Amazingly, the tests all still passed because some of the input columns
and unrestrictive (field) types.  Have added tests which do fail with
this error.

* Fix broken tests

Since validation for valid tests is now enabled, it meant some existing
tests were no longer well-formed.

* more fixes for tests