feat: support defensive padding (#618)
This adds support for defensive padding, and enables it by default.
Defensive padding is the conversative prepending of padding rows to
ensure that constraints in the active region are not clipped.  This
reflects the fact that such constraints may be subject to padding by the
prover anyway and, hence, we simply get ahead of the curve by doing
adding it automatically.

This also adds support to the trace command to expand the trace prior to
filtering, etc and also to allow defensive padding and/or spillage to be
applied.