feat: support perspectives in inspector (#631)
* support attributes in binfile format

This adds support for arbitrary attributes in the binfile format.  Such
attributes can be used, for example, for encoding "source maps" which
map source columns to their allocated columns, etc.

* support source column mapping

This introduces the notion of a "source map" which (at this time) simply
maps source-level columns to their respective registers.

* inspector now reading debug metadata

The inspector is now reading the debug metadata, and display columns
correctly with respect to perspectives.  However, its not yet hiding
perspectives.

* refactored module view

This puts through a fairly large refactoring of the module view which
helps to separate out state from the view itself, and also simplifies
the logic in many places.

* more refactoring

At this stage, we are almost now ready to actually get the perspectives
highlighting working.

* support perspective filtering

This hides values from cells in a given perspective when that cell is
not active (i.e. because the perspective is not active).

* fix compilation failure

Some invalid tests were crashing rather than failing gracefully.