stratify declaration finalisation
This stratifies the way that declarations are processed in order to
allow assignments to be resolved prior to constraints.  This, in turn,
allows constraints in the global module to refer to columns in
submodules, etc.

The stratification relies on there being essentially two classes of
declaration.  First we have those which declare (or define) one or more
columns (this includes e.g. defcolumns but also defpermutation, etc).
Secondly, we have those which simply refer to existing columns (e.g.
constraints, functions, etc).  The key is that, as it stands, these are
essentially separate groups.

In fact, the two groups are not fully separate in the case of
perspectives.  Specifically, in this case, it is technically permitted
to give a selector expression.  Such an expression, therefore, could
presumably attempt to refer to a column defined in another module, and
that would cause problems.  However, at this time, selectors are never
used in this way.