eliminate catchall "cyclic declaration" error
This eliminates the dreaded "cyclic declaration" error and, instead,
reports specific errors on symbols which could not be resolved.  This
makes the error more targetted and, hence, more helpful.  However, the
errors reported could still be improved if we were able to determine
what kind of symbol we were looking for at the given point.