feat: add initial coverage command (#684)
* feat: add initial coverage command

This initial command is fairly rough, but it is digging out suitable
data.  As such, it provides a gateway to understand the coverage
results.

* add `Constraint.Contexts()`

This methods provides generic access to the enclosing contexts of a
given constraint.  Whilst, in most cases, there is exactly one enclosing
context for a constraint, there are constraints (lookups) which have
more than one.

* improve coverage reports / format

This improves the coverage reports by grouping things by module.  This
required a few surprising changes to the definition of a Constraint.
More is needed to determine the total number of branches with a
constraint.

This also puts in place a more flesxibile infrastructure for producing
coverage reports.  For example, we can now filter using a regular
expression.  However, there is still quite a way to go.