feat: support batched mode in trace command (#694)
* support batched mode in trace command

This adds a batched mode to the trace command.  This is helpful because
it allows us to expand a given batch of traces, such as for the purposes
of generating useful tests, etc.

* add expanded tests

This adds a number of expanded trace tests for inverses, permutations,
lookups and range constraints.  This helps to check that the necessary
checks are being performed on the expanded trace.  However, whilst this
provides a good starting point, there is certainly still work to be done
here.