feat: lexicograpgic ordering constraint (#715)
* implemented `defsorted` construct

This puts through the implementation of the `defsorted` construct which,
under-the-hood, applies sorting constraints onto a given set of columns.
The sorting constraints are applied lexicographically according to
designated signs (i.e. sort directions) given by the user.

* extend testing for defsorted

This adds a reasonably large number of relatively extensive tests for
the defsorted construct.  As such, a small number of bugs are fixed.
Also, invalid cases are added to catch syntax errors.