feat: strict lexicographic ordering constraint (#723)
* parser now accepting defstrictsorted

* Support strict sorting at HIR / MIR level

This adds support for the strict sorting constraint the HIR and MIR
levels.  However, it remains to update the gadget at the AIR level.

* support single column strict sorts

This puts in place necessary logic for managing single column strict
sorts.

* support strict lex sort for > 1 columns

This adds support for a strict lexicographic sort in the multicolumn
case.  In particular, the lexiocgraphic sorting gadget is now
parameterised to support a strictness flag, along with a selector
column.

* add accept tests for strictsorted constraint

This adds a range of automatically generated accept tests for the new
strictly sorted constraints.

* add reject tests for strict lexicographic sorts

This adds a number of reject tests for the strictsorted declaration.

* fix sorted tests

Two sorted tests where ammended, but incorrectly. Fixed now.