feat: support constant types (#730)
* feat: support constant types

This adds support for explicit types being given to constants.  The
primary purpose for this, currently, is to help direct generation of the
Trace API.  For example, we can specify that a given constant should
have a given type and, hence, ensure the generate type in the target
language (e.g. Java) is large enough to hold any value of this type.

* bug fix for `UintType.IntBound()`