feat: support externalised constants (#751)
Adds support for "externalised constants".  These are normal constants declared
with the `:extern` attribute.  This means they can be changed int the `zkevm.bin`
 during checking, etc.  Some restrictions are currently placed on them.

This introduces the notion of a "const context" in which one cannot, for
example, refer to an externalised constant.  This prohibits some complex
scenarious where externalise constants are compiled away.