feat: support improved register allocation (#756)
This adds a preliminary implementation of an advanced register allocator which merges columns that are not proven, but have different types.  At this stage, however, the allocator is not enabled as, currently, this is unsafe.  Further improvements to the register allocation algorithm are required before it can be enabled by default.

In addition, this streamlines the configuration for compiling `corset` constraints down to HIR.