implement mir.Equation
This adds support for equations at the MIR level.  This is the first
step towards eliminating normalisation expressions, however there is a
way to go.

This puts in place incomplete support for handling conditions at the
corset level.  The idea is to manage this through the if-condition since
this is never exposed directly to the user.