feat: reject columns no type (#601) (#758)
This imposes a strong restriction on the constraints that all user-defined columns must be given explicit types.  This is a pre-step towards field agnosticity as, with this feature, field types will no longer make sense at the source level.  Computed columns are not required (or even allowed) to have explicit types at this stage.  However, in the future, we could either: (i) demand that they have explicit types; or (ii) allow them to have explicit types which are enforced.  

As a result of this feature, a larger number of the existing tests were updated to meet the stricter requirement.  A very small number are left which still use the field type (albeit this is now given explicitly).  This is done to simplify the transition whilst we figure out how to migrate these particular tests.  To make this work, we introduced a notion of "strictness".  Specifically, strict mode is the default and this prohibits using the field type; but, the tests are run in non-strict mode which allows the field type to be used (provided this is explicit).

---------

Co-authored-by: amkCha <29160563+amkCha@users.noreply.github.com>